CHanged health to floats and removed Jump

+++ Assets/Scripts/Combat_Refactored/HitBox.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class HitBox : MonoBehaviour
 {
-    public int amount;
+    public float amount;
     public BoxCollider2D hitBoxCollider;
 
     private void Awake()
